+++ webui/helpers.go
@@ -274,9 +274,9 @@ func uptimeInDays(req *http.Request) string {
 	return fmt.Sprintf("%.0f", time.Since(ctx(req).Server().Stats.StartedAt).Seconds()/float64(86400))
 }
 
-func displayRss(rssKb int) string {
+func displayRss(rssKb int64) string {
 	if rssKb < 100000 {
-		return strconv.FormatInt(int64(rssKb), 10) + " KB"
+		return strconv.FormatInt(rssKb, 10) + " KB"
 	} else if rssKb < 10000000 {
 		return strconv.FormatFloat(float64(rssKb)/1024, 'f', 1, 64) + " MB"
 	} else {
